Update documentation-guidelines.html.md
This requirement seems ... a bit silly:
```
International/British English for Text, USA/American English for Code

We love it when people fix spelling errors and typos! We use International/British English for text (only 27% of our visitors are from the USA), and USA/American English for code. Code examples should (where it makes sense to) comply with the Bevry Coding Standards.
```
I changed it a bit.